Renamed and moved the interface file for surface query service
